Fixes clean architecture dependency graph
