+++ .idea/gradle.xml
@@ -9,7 +9,6 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/starwars-core" />
             <option value="$PROJECT_DIR$/starwars-data" />
             <option value="$PROJECT_DIR$/starwars-domain" />
           </set>
