+++ starwars-data/build.gradle
@@ -43,10 +43,10 @@ android {
             returnDefaultValues = true
         }
     }
-
-    dataBinding {
-        enabled = true
-    }
+//
+//    dataBinding {
+//        enabled = true
+//    }
 
 }
 
@@ -76,7 +76,7 @@ androidExtensions {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation project(":starwars-core")
+    implementation project(":starwars-domain")
 
     implementation baseDependencies.kotlin_std_lib
 //    implementation baseDependencies.anko
