+++ starwars-domain/src/main/kotlin/br/com/devhernand/starwars/domain/api/br/com/devhernand/starwars/domain/main/MainUseCaseImpl.kt
@@ -1,7 +1,4 @@
 package br.com.devhernand.starwars.domain.api.br.com.devhernand.starwars.domain.main
-
-import br.com.devhernand.starwars.core.Product
-import br.com.devhernand.starwars.core.ProductRepository
 import io.reactivex.Single
 
 class MainUseCaseImpl(val productRepository: ProductRepository) : MainUseCase {
